Add files via upload
The Topology shown is the infrastructure Topology. 

This type of Topology is present in my own home. The computers and printers in my home have to go through the gateway( router) first in order to perform a simple function such as printing documents. The router gives the signal to both the computer and the printer. The computer sends the command to printer the printer then goes back to the printer , the router sends the signal to the computer and soon you have a document coming out of the printer. The same thing happens with the tablet, the tablet sends the signal to the printer, the router is wireless and therefore emits a wireless signal and once that signal reaches the printer , a document is printed out.